lesson 1

+++ lesson1/task2.cpp
@@ -5,7 +5,7 @@ int main() {
     setlocale(LC_ALL, "Russian");
     cout << "Выполнил Кравченко Алексей 9КС-232" << endl;
     
-    double a, b;
+    int a, b;
     cout << "Введите два числа: " << endl;
     cin >> a >> b;
